+++ Zeebe.Worker/Models/Command/CreateInstanceRequest.cs
@@ -5,6 +5,6 @@ namespace Zeebe.Worker.Models.Command
     public record CreateInstanceRequest(
         string BpmnProcessId,
         long? WorkflowKey,
-        short? Version,
+        int? Version,
         Dictionary<string, string> Variables);
 }
