+++ Zeebe.Worker/Models/Command/CreateInstanceResponse.cs
@@ -3,6 +3,6 @@ namespace Zeebe.Worker.Models.Command
     public record CreateInstanceResponse(
         long? WorkflowKey,
         string BpmnProcessId,
-        short? Version,
+        int? Version,
         long? WorkflowInstanceKey);
 }
