+++ Zeebe.Worker/Models/Command/FailJobRequest.cs
@@ -4,6 +4,6 @@ namespace Zeebe.Worker.Models.Command
 {
     public record FailJobRequest(
         [Required] long? JobKey,
-        [Required] short? Retries,
+        [Required] int? Retries,
         string ErrorMessage);
 }
